% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{eda}
\usage{
eda(model_df)
}
\arguments{
\item{model_df}{Data frame containing potential variables from the dataset}
}
\value{
Diagram matrices containing scatterplot between variables (at lower triangle of the matrix),
histogram and name of certain variables (at diagonal), as well as correlation coefficients (at upper triangle of the matrix)
}
\description{
Exploratory Data Analysis of the selected variables from linear model
}
\examples{
## SLR (Simple Linear Regression)
model_SLR = lm_mat("mpg", c("cyl"), mtcars, beta0 = TRUE)
eda(mtcars[,c(model_SLR$yvar, model_SLR$xvar)])

## MLR (Multiple Linear Regression with intercept)
model_MLR1 = lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = TRUE))
eda(mtcars[,c(model_MLR1$yvar, model_MLR1$xvar)])

## MLR (Multiple Linear Regression without intercept)
model_MLR2 = lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = FALSE))
eda(mtcars[,c(model_MLR2$yvar, model_MLR2$xvar)])

}
