% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef1}
\alias{coef1}
\title{coef1}
\usage{
coef1(model)
}
\arguments{
\item{model}{the list derived from lm_mat() function}
}
\value{
coefficient table of the linear model, R squared and Adjusted R squared,
as well as covariance matrix of the model
}
\description{
Parameter estimation and inference of the linear model
}
\examples{
## SLR (Simple Linear Regression)
model_SLR = lm_mat("mpg", c("cyl"), mtcars, beta0 = TRUE)
coef1(model_SLR)

## MLR (Multiple Linear Regression with intercept)
model_MLR1 = lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = TRUE)
coef1(model_MLR1)

## MLR (Multiple Linear Regression without intercept)
model_MLR2 = lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = FALSE)
coef1(model_MLR2)

}
