% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mat.R
\name{lm_mat}
\alias{lm_mat}
\title{lm_mat}
\usage{
lm_mat(Y_lab, X_lab, data, beta0 = TRUE)
}
\arguments{
\item{Y_lab}{the name of the outcome(Y) variable (characters)}

\item{X_lab}{the name(s) of the predictor(s) (characters)}

\item{data}{data frame from which the variables are selected}

\item{beta0}{Intercept should be included if `beta0`=`TRUE`, or not if `beta0`=`FALSE`}
}
\value{
basic information from Y and X, as well as matrix notations frequently used
}
\description{
Gets the matrix notation and basic information
from outcome(Y) and predictors(X) constructing a linear model
}
\examples{
## SLR (Simple Linear Regression)
lm_mat("mpg", c("cyl"), mtcars, beta0 = TRUE)

## MLR (Multiple Linear Regression with intercept)
lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = TRUE))
lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = TRUE))

## MLR (Multiple Linear Regression without intercept)
lm_mat("mpg", c("cyl", "disp"), mtcars, beta0 = FALSE))

}
